# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Halls Head Cricket Club Stats API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in as an admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentAdmin
      tags: [auth]
      summary: Get the currently signed-in admin
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Not signed in
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List all admin users
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create a new admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Created admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /admins/{id}:
    patch:
      operationId: updateAdmin
      tags: [admins]
      summary: Update an admin (display name, username, password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdate"
      responses:
        "200":
          description: Updated admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "404":
          description: Not found
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Delete an admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          description: Cannot delete the last remaining admin
        "404":
          description: Not found
  /honour-boards:
    get:
      operationId: listHonourBoards
      tags: [honourBoards]
      summary: List all honour board definitions
      responses:
        "200":
          description: Honour board definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HonourBoard"
    post:
      operationId: createHonourBoard
      tags: [honourBoards]
      summary: Create a honour board definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HonourBoardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HonourBoard"
  /honour-boards/{key}:
    patch:
      operationId: updateHonourBoard
      tags: [honourBoards]
      summary: Update an honour board definition
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HonourBoardUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HonourBoard"
        "404":
          description: Not found
    delete:
      operationId: deleteHonourBoard
      tags: [honourBoards]
      summary: Soft-delete an honour board
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /honour-boards/{key}/overrides:
    get:
      operationId: listHonourBoardOverrides
      tags: [honourBoards]
      summary: List overrides for a single honour board
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Override rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HonourBoardOverride"
    post:
      operationId: upsertHonourBoardOverride
      tags: [honourBoards]
      summary: Create or update an override for (board, player)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HonourBoardOverrideInput"
      responses:
        "200":
          description: Upserted override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HonourBoardOverride"
  /honour-boards/{key}/overrides/{playerId}:
    delete:
      operationId: deleteHonourBoardOverride
      tags: [honourBoards]
      summary: Remove an override
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    Admin:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
    AdminInput:
      type: object
      required: [username, displayName, password]
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
    AdminUpdate:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    HonourBoard:
      type: object
      required: [id, key, label, title, subtitle, headlineLabel, supportingLabel, displayOrder]
      properties:
        id:
          type: integer
        key:
          type: string
        label:
          type: string
        title:
          type: string
        subtitle:
          type: string
        headlineLabel:
          type: string
        supportingLabel:
          type: string
        displayOrder:
          type: integer
    HonourBoardInput:
      type: object
      required: [key, label, title]
      properties:
        key:
          type: string
        label:
          type: string
        title:
          type: string
        subtitle:
          type: string
        headlineLabel:
          type: string
        supportingLabel:
          type: string
        displayOrder:
          type: integer
    HonourBoardUpdate:
      type: object
      properties:
        label:
          type: string
        title:
          type: string
        subtitle:
          type: string
        headlineLabel:
          type: string
        supportingLabel:
          type: string
        displayOrder:
          type: integer
    HonourBoardOverride:
      type: object
      required: [id, boardKey, playerId, pinned, hidden, note]
      properties:
        id:
          type: integer
        boardKey:
          type: string
        playerId:
          type: integer
        pinned:
          type: boolean
        hidden:
          type: boolean
        note:
          type: string
    HonourBoardOverrideInput:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: integer
        pinned:
          type: boolean
        hidden:
          type: boolean
        note:
          type: string