# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halls Head Cricket Club Stats API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  username: string;
  displayName: string;
  createdAt: string;
}