# lib/api-zod/src/generated/types/honourBoardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halls Head Cricket Club Stats API
 * OpenAPI spec version: 0.1.0
 */

export interface HonourBoardInput {
  key: string;
  label: string;
  title: string;
  subtitle?: string;
  headlineLabel?: string;
  supportingLabel?: string;
  displayOrder?: number;
}