# lib/api-zod/src/generated/types/honourBoardUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Halls Head Cricket Club Stats API
 * OpenAPI spec version: 0.1.0
 */

export interface HonourBoardUpdate {
  label?: string;
  title?: string;
  subtitle?: string;
  headlineLabel?: string;
  supportingLabel?: string;
  displayOrder?: number;
}